+++ relays/bin-substrate/src/bridges/kusama_polkadot/polkadot_headers_to_bridge_hub_kusama.rs
@@ -21,8 +21,8 @@ use crate::cli::bridge::{CliBridgeBase, RelayToRelayHeadersCliBridge};
 use async_trait::async_trait;
 use relay_substrate_client::{AccountKeyPairOf, Client};
 use substrate_relay_helper::{
-	finality::{engine::Grandpa as GrandpaFinalityEngine, SubstrateFinalitySyncPipeline},
-	TransactionParams,
+	finality::SubstrateFinalitySyncPipeline,
+	finality_base::engine::Grandpa as GrandpaFinalityEngine, TransactionParams,
 };
 
 /// Description of Polkadot -> KusamaBridgeHub finalized headers bridge.
