+++ relays/bin-substrate/src/bridges/rialto_parachain_millau/millau_headers_to_rialto_parachain.rs
@@ -35,8 +35,9 @@
 //! Millau-to-RialtoParachain headers sync entrypoint.
 
 use crate::cli::bridge::{CliBridgeBase, MessagesCliBridge, RelayToRelayHeadersCliBridge};
-use substrate_relay_helper::finality::{
-	engine::Grandpa as GrandpaFinalityEngine, SubstrateFinalitySyncPipeline,
+use substrate_relay_helper::{
+	finality::SubstrateFinalitySyncPipeline,
+	finality_base::engine::Grandpa as GrandpaFinalityEngine,
 };
 
 substrate_relay_helper::generate_submit_finality_proof_call_builder!(
