+++ relays/bin-substrate/src/bridges/rococo_wococo/rococo_headers_to_bridge_hub_wococo.rs
@@ -21,8 +21,8 @@ use crate::cli::bridge::{CliBridgeBase, RelayToRelayHeadersCliBridge};
 use async_trait::async_trait;
 use relay_substrate_client::{AccountKeyPairOf, Client};
 use substrate_relay_helper::{
-	finality::{engine::Grandpa as GrandpaFinalityEngine, SubstrateFinalitySyncPipeline},
-	TransactionParams,
+	finality::SubstrateFinalitySyncPipeline,
+	finality_base::engine::Grandpa as GrandpaFinalityEngine, TransactionParams,
 };
 
 /// Description of Rococo -> Wococo finalized headers bridge.
