+++ relays/bin-substrate/src/bridges/westend_millau/westend_headers_to_millau.rs
@@ -17,9 +17,9 @@
 //! Westend-to-Millau headers sync entrypoint.
 
 use crate::cli::bridge::{CliBridgeBase, RelayToRelayHeadersCliBridge};
-use substrate_relay_helper::finality::{
-	engine::Grandpa as GrandpaFinalityEngine, DirectSubmitGrandpaFinalityProofCallBuilder,
-	SubstrateFinalitySyncPipeline,
+use substrate_relay_helper::{
+	finality::{DirectSubmitGrandpaFinalityProofCallBuilder, SubstrateFinalitySyncPipeline},
+	finality_base::engine::Grandpa as GrandpaFinalityEngine,
 };
 
 /// Description of Westend -> Millau finalized headers bridge.
